Add Rekognition Labels to autosuggest fields
Since we introduced Elasticsearch to AJAX requests the expected behaviour of the typeahead / autocomplete search in the media library for instance has not been returning results until a full word is entered which was confusing to users and devs alike.

With the changes in https://github.com/humanmade/altis-enhanced-search/pull/92 this PR will add the rekognition labels to the mapping to be analysed using ngrams and adds them to the search in media library AJAX requests.